% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_gam.R
\name{logLik_gam}
\alias{logLik_gam}
\title{logLik_gam}
\usage{
logLik_gam(model)
}
\arguments{
\item{model}{detail}
}
\description{
Log likelihood function for use in model selection of mixed models using gam or bam
Here's a function for extracting the log likelihood and (harder) the degrees of freedom
out of a gam or bam model.  The degrees of freedom are the number of fixed effects plus
the number of variances (not the edf as is usually reported by gam) and assumes the
models are fitted by ML.
}
\details{
The function has been tested with a range of smoothers, but as good check would involve
fitting models in gam and also in gamm4 and check the df are the same.
}
