% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gamm4.R
\name{my_gamm4}
\alias{my_gamm4}
\title{fixes bug in gamm4 to allow different optimisers}
\usage{
my_gamm4(
  formula,
  random = NULL,
  family = gaussian(),
  data = list(),
  weights = NULL,
  subset = NULL,
  na.action,
  knots = NULL,
  drop.unused.levels = TRUE,
  REML = TRUE,
  control = NULL,
  start = NULL,
  verbose = 0L,
  mer_only = FALSE,
  ...
)
}
\arguments{
\item{formula}{A GAM formula (see also \link{formula.gam} and
\link{gam.models}). This is like the formula for a glm except that
smooth terms (s and t2 but not te) can be added to the right hand
side of the formula. Note that ids for smooths and fixed smoothing
parameters are not supported.ils}

\item{random}{An optional formula specifying the random effects structure in lmer style. See example \link{gamm4}.}

\item{family}{A family as used in a call to glm or gam.}

\item{data}{A data frame or list containing the model response variable and covariates required by the formula. By default the variables are taken from environment(formula), typically the environment from which gamm4 is called.}

\item{weights}{a vector of prior weights on the observations. NULL is equivalent to a vector of 1s. Used, in particular, to supply the number-of-trials for binomial data, when the response is proportion of successes.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain ‘NA’s. The default is set by the ‘na.action’ setting of ‘options’, and is ‘na.fail’ if that is unset. The “factory-fresh” default is ‘na.omit’.}

\item{knots}{this is an optional list containing user specified knot values to be used for basis construction. Different terms can use different numbers of knots, unless they share a covariate.}

\item{drop.unused.levels}{by default unused levels are dropped from factors before fitting. For some smooths involving factor variables you might want to turn this off. Only do so if you know what you are doing.}

\item{REML}{passed on to lmer fitting routines (but not glmer fitting routines) to control whether REML or ML is used.}

\item{control}{lmerControl or glmerControl list as appropriate (NULL means defaults are used).}

\item{start}{starting value list as used by lmer or glmer.}

\item{verbose}{passed on to fitting lme4 fitting routines.}

\item{mer_only}{if TRUE returns the \code{mer} part of the out and does
not go on to fit the gam object. Useful for speeding up model
selection and when interest lies in the log-likelihood of the model.}

\item{...}{additional arguments to gamm4}
}
\value{
object of class gamm4, but with an aditional call object (like lm,
glm etc.) for use in model selection
}
\description{
fixes bug in gamm4 to allow different optimisers
}
\details{
fixes bug in gamm4 to allow different optimisers
based on gamm4 0.2-5
need to explicitly pass optimiser to optimizeLmer or (twice) to optimizeGlmer,
twice since the latter optimisation is a two-stage process
}
\seealso{
gamm4
}
